+++ smart-contracts/contracts/v1/utils/MathUtils.sol
@@ -1,12 +1,16 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.15;
 
-function getRandomInRange(uint min, uint max, uint nonce) view returns (uint) {
-    uint rand = _rand(nonce);
-    uint range = max - min + 1;
+function getRandomInRange(
+    uint256 min,
+    uint256 max,
+    uint256 nonce
+) view returns (uint256) {
+    uint256 rand = _rand(nonce);
+    uint256 range = max - min + 1;
     return (rand % range) + min;
 }
 
-function _rand(uint nonce) view returns (uint) {
-    return uint(keccak256(abi.encodePacked(block.timestamp, block.difficulty, tx.origin, nonce)));
+function _rand(uint256 nonce) view returns (uint256) {
+    return uint256(keccak256(abi.encodePacked(block.timestamp, block.difficulty, tx.origin, nonce)));
 }
